Add option.extendDefaultFields
The method could be used to store additional data from session to the session
table, for example userId of the user.

The main reason I needed this functionality is to be able to select all
sessions by certain user. For example, I added `userId` string to
session model, and I want that field to be read from session.

The way to do it is to implement an extendDefaultFields function:

```
function extend(defaults, session) {
    return {
        data: defaults.data,
        expires: defaults.expires,
        userId: session.userId
    };
}

store = new SessionStore({db: db, extendDefaultFields: extend});
```

and now userId will be stored in the database.